+++ config.env.php
@@ -9,5 +9,6 @@ define('RMQ_VHOST', getenv('RMQ_VHOST'));
 define('RMQ_USER', getenv('RMQ_USER'));
 define('RMQ_PASS', getenv('RMQ_PASS'));
 define('RMQ_QUEUE_IN', getenv('RMQ_QUEUE_IN'));
+define('RMQ_QUEUE_LOG', getenv('RMQ_QUEUE_LOG'));
 define('RMQ_RECONNECT_TIMEOUT', getenv('RMQ_RECONNECT_TIMEOUT'));
 define('RMQ_TICK_TIMEOUT', getenv('RMQ_TICK_TIMEOUT'));