+++ start.php
@@ -50,7 +50,8 @@ $camundaConfig = [
 $rmqConfig = [
     'queue'            => RMQ_QUEUE_IN,
     'tickTimeout'      => RMQ_TICK_TIMEOUT,
-    'reconnectTimeout' => RMQ_RECONNECT_TIMEOUT
+    'reconnectTimeout' => RMQ_RECONNECT_TIMEOUT,
+    'queueLog'         => RMQ_QUEUE_LOG
 ];
 
 // Run worker