+++ CHANGELOG.md
@@ -5,13 +5,25 @@ BPM listener on PHP. Using to catch message from queue and resume business proce
 ## [Unreleased]
 
 ### Planned
+
 - Nothing
+
+## [0.2] - 2020-01-24
+
+### Changed 
+
+- Version kubia/camunda from 0.2 to 0.3
+
+### Added
+
+- Logging to elasticsearch
 
 ## [0.1] - 2020-01-06
 
 ### Added
 
-- Nothing
+- First worked version
 
-[unreleased]: https://gitlab.com/quancy-core/bpm-catcher/-/tags/0.1
-[0.1]: https://gitlab.com/quancy-core/bpm-catcher/-/tags/0.1
+[unreleased]: https://gitlab.com/quancy-core/bpm-listener/-/tags/0.2
+[0.2]: https://gitlab.com/quancy-core/bpm-listener/-/tags/0.2
+[0.1]: https://gitlab.com/quancy-core/bpm-listener/-/tags/0.1